refactor journey api & some layout fixes

+++ lib/api/journey.ts
@@ -3,8 +3,20 @@ import { createClient } from "../supabase/client";
 
 const supabase = createClient();
 
-export const fetchJourneys = async () => {
-  const { data: journeys, error } = await supabase.from("journeys").select("*");
+export const fetchUserJourneys = async () => {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  const userId = user?.id;
+
+  if (!userId) throw new Error("No logged in user.");
+
+  const { data: journeys, error } = await supabase
+    .from("journeys")
+    .select("*")
+    .eq("user_id", userId)
+    .order("updated_at", { ascending: false });
 
   if (error) {
     throw new Error(errorHandler(error));
@@ -26,11 +38,18 @@ export const fetchJourneyById = async (id: string) => {
 };
 
 export const createJourney = async (
-  userId: string,
   name: string,
   description?: string,
   coverImgUrl?: string
 ) => {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  const userId = user?.id;
+
+  if (!userId) throw new Error("No logged in user.");
+
   const { data, error } = await supabase
     .from("journeys")
     .insert([
@@ -56,12 +75,21 @@ export const editJourneyById = async (
   description?: string,
   coverImgUrl?: string
 ) => {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  const userId = user?.id;
+
+  if (!userId) throw new Error("No logged in user.");
+
   const { data, error } = await supabase
     .from("journeys")
     .update({
       name,
       description,
       cover_img_url: coverImgUrl,
+      updated_at: new Date().toISOString(),
     })
     .eq("id", journeyId)
     .select();
@@ -74,6 +102,14 @@ export const editJourneyById = async (
 };
 
 export const deleteJourneyById = async (id: string) => {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  const userId = user?.id;
+
+  if (!userId) throw new Error("No logged in user.");
+
   const { error } = await supabase.from("journeys").delete().eq("id", id);
 
   if (error) {
